feat: structure change
