+++ lib/src/config/routes/routes.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:news_app/src/features/daily_news/presentation/pages/home/account_screen.dart';
 
 import '../../features/daily_news/presentation/pages/article_detail/article_screen.dart';
-import '../../features/daily_news/presentation/pages/home/discover_screen.dart';
+import '../../features/daily_news/presentation/pages/discover/discover_screen.dart';
 import '../../features/daily_news/presentation/pages/home/home_screen.dart';
 
 class AppRoutes {
