+++ lib/src/features/daily_news/presentation/pages/article_detail/article_screen.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:news_app/src/features/daily_news/domain/entities/article.dart';
 
-import '../../widgets/home/custom_tag.dart';
-import '../../widgets/home/image_container.dart';
+import '../../../../../core/components/image_container.dart';
+import 'components/news_body.dart';
+import 'components/news_headline.dart';
 
 class ArticleScreen extends StatelessWidget {
   const ArticleScreen({Key? key}) : super(key: key);
@@ -15,6 +16,7 @@ class ArticleScreen extends StatelessWidget {
     return ImageContainer(
       width: double.infinity,
       imageUrl: article.urlToImage ?? '',
+      wholeBlur: true,
       child: Scaffold(
         backgroundColor: Colors.transparent,
         appBar: AppBar(
@@ -24,164 +26,11 @@ class ArticleScreen extends StatelessWidget {
         ),
         extendBodyBehindAppBar: true,
         body: ListView(
-          children: [_NewsHeadline(article: article), _NewsBody(article: article)],
+          children: [
+            NewsHeadline(article: article),
+            NewsBody(article: article),
+          ],
         ),
-      ),
-    );
-  }
-}
-
-class _NewsBody extends StatelessWidget {
-  const _NewsBody({
-    Key? key,
-    required this.article,
-  }) : super(key: key);
-
-  final ArticleEntity article;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(20.0),
-      decoration: const BoxDecoration(
-        borderRadius: BorderRadius.only(
-          topLeft: Radius.circular(20.0),
-          topRight: Radius.circular(20.0),
-        ),
-        color: Colors.white,
-      ),
-      child: Column(
-        children: [
-          Row(
-            children: [
-              CustomTag(
-                backgroundColor: Colors.black,
-                children: [
-                  CircleAvatar(
-                    radius: 10,
-                    backgroundImage: NetworkImage(
-                      article.urlToImage ?? '',
-                    ),
-                  ),
-                  const SizedBox(width: 10),
-                  SizedBox(
-                    width: MediaQuery.sizeOf(context).width / 2,
-                    child: Text(
-                      article.author!,
-                      maxLines: 2,
-                      overflow: TextOverflow.ellipsis,
-                      style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                            color: Colors.white,
-                          ),
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(width: 10),
-              CustomTag(
-                backgroundColor: Colors.grey.shade200,
-                children: [
-                  const Icon(
-                    Icons.timer,
-                    color: Colors.grey,
-                  ),
-                  const SizedBox(width: 10),
-                  Text('${DateTime.now().difference(article.publishedAt!).inHours}h', style: Theme.of(context).textTheme.bodyMedium),
-                ],
-              ),
-              // const SizedBox(width: 10),
-              // CustomTag(
-              //   backgroundColor: Colors.grey.shade200,
-              //   children: [
-              //     const Icon(
-              //       Icons.remove_red_eye,
-              //       color: Colors.grey,
-              //     ),
-              //     const SizedBox(width: 10),
-              //     Text(
-              //       '${article.views}',
-              //       style: Theme.of(context).textTheme.bodyMedium,
-              //     ),
-              //   ],
-              // ),
-            ],
-          ),
-          const SizedBox(height: 20),
-          Text(
-            article.title ?? 'Title',
-            style: Theme.of(context).textTheme.headlineSmall!.copyWith(fontWeight: FontWeight.bold),
-          ),
-          const SizedBox(height: 20),
-          Text(
-            article.content!,
-            style: Theme.of(context).textTheme.bodyMedium!.copyWith(height: 1.5),
-          ),
-          const SizedBox(height: 20),
-          GridView.builder(
-              shrinkWrap: true,
-              itemCount: 2,
-              gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                crossAxisCount: 2,
-                childAspectRatio: 1.25,
-              ),
-              itemBuilder: (context, index) {
-                return ImageContainer(
-                  width: MediaQuery.of(context).size.width * 0.42,
-                  imageUrl: article.urlToImage ?? '',
-                  margin: const EdgeInsets.only(right: 5.0, bottom: 5.0),
-                );
-              })
-        ],
-      ),
-    );
-  }
-}
-
-class _NewsHeadline extends StatelessWidget {
-  const _NewsHeadline({
-    Key? key,
-    required this.article,
-  }) : super(key: key);
-
-  final ArticleEntity article;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(20.0),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.end,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          SizedBox(
-            height: MediaQuery.of(context).size.height * 0.15,
-          ),
-          // CustomTag(
-          //   backgroundColor: Colors.grey.withAlpha(150),
-          //   children: [
-          //     Text(
-          //       article.category,
-          //       style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-          //         color: Colors.white,
-          //       ),
-          //     ),
-          //   ],
-          // ),
-          // const SizedBox(height: 10),
-          Text(
-            article.title ?? 'Title',
-            style: Theme.of(context).textTheme.headlineSmall!.copyWith(
-                  fontWeight: FontWeight.bold,
-                  color: Colors.white,
-                  height: 1.25,
-                ),
-          ),
-          const SizedBox(height: 10),
-          Text(
-            article.description ?? '',
-            style: Theme.of(context).textTheme.bodyMedium!.copyWith(color: Colors.white),
-          ),
-        ],
       ),
     );
   }
