+++ lib/src/features/daily_news/presentation/pages/discover/discover_screen.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+
+import '../../../../../core/components/bottom_nav_bar.dart';
+import 'components/category_news.dart';
+import 'components/discover_news.dart';
+
+class DiscoverScreen extends StatelessWidget {
+  const DiscoverScreen({Key? key}) : super(key: key);
+
+  static const routeName = '/discover';
+
+  @override
+  Widget build(BuildContext context) {
+    List<String> tabs = ['Health', 'Politics', 'Art', 'Food', 'Science'];
+
+    return DefaultTabController(
+      initialIndex: 0,
+      length: tabs.length,
+      child: Scaffold(
+        appBar: AppBar(
+          backgroundColor: Colors.transparent,
+          elevation: 0,
+          leading: IconButton(
+            onPressed: () {},
+            icon: const Icon(
+              Icons.menu,
+              color: Colors.black,
+            ),
+          ),
+        ),
+        bottomNavigationBar: const BottomNavBar(index: 1),
+        body: ListView(
+          padding: const EdgeInsets.all(20.0),
+          children: [
+            const DiscoverNews(),
+            CategoryNews(tabs: tabs),
+          ],
+        ),
+      ),
+    );
+  }
+}
