+++ lib/src/features/daily_news/presentation/pages/home/home_screen.dart
@@ -1,17 +1,30 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:news_app/src/features/daily_news/presentation/bloc/article/remote/remote_article_bloc.dart';
-import 'package:news_app/src/features/daily_news/presentation/bloc/article/remote/remote_article_state.dart';
+import 'package:news_app/src/features/daily_news/presentation/pages/home/components/breaking_news.dart';
 
-import '../../widgets/home/bottom_nav_bar.dart';
-import '../../widgets/home/custom_tag.dart';
-import '../../widgets/home/image_container.dart';
-import '../article_detail/article_screen.dart';
+import '../../../../../core/components/bottom_nav_bar.dart';
+import '../../../../../injection_container.dart';
+import '../../../domain/enums/news_category.dart';
+import '../../bloc/article/remote/remote_article_event.dart';
+import 'components/news_of_day.dart';
 
 class HomeScreen extends StatelessWidget {
-  const HomeScreen({Key? key}) : super(key: key);
+  const HomeScreen({super.key});
 
   static const routeName = '/';
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider<RemoteArticlesBloc>(
+      create: (context) => sl()..add(GetArticles(category: NewsCategory.general.category)),
+      child: const _HomeScreen(),
+    );
+  }
+}
+
+class _HomeScreen extends StatelessWidget {
+  const _HomeScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -32,164 +45,10 @@ class HomeScreen extends StatelessWidget {
       body: ListView(
         padding: EdgeInsets.zero,
         children: const [
-          _NewsOfTheDay(),
-          _BreakingNews(),
+          NewsOfTheDay(),
+          BreakingNews(),
         ],
       ),
-    );
-  }
-}
-
-class _BreakingNews extends StatelessWidget {
-  const _BreakingNews({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<RemoteArticlesBloc, RemoteArticleState>(
-      builder: (_, state) {
-        if (state is RemoteArticlesLoading) {
-          return const Center(
-            child: CircularProgressIndicator(),
-          );
-        } else {
-          return Padding(
-            padding: const EdgeInsets.all(20.0),
-            child: Column(
-              children: [
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Text(
-                      'Breaking News',
-                      style: Theme.of(context).textTheme.headlineSmall!.copyWith(fontWeight: FontWeight.bold),
-                    ),
-                    Text('More', style: Theme.of(context).textTheme.bodyLarge),
-                  ],
-                ),
-                const SizedBox(height: 20),
-                SizedBox(
-                  height: 250,
-                  child: ListView.builder(
-                    scrollDirection: Axis.horizontal,
-                    itemCount: state.articles!.length,
-                    itemBuilder: (context, index) {
-                      return Container(
-                        width: MediaQuery.of(context).size.width * 0.5,
-                        margin: const EdgeInsets.only(right: 10),
-                        child: InkWell(
-                          onTap: () {
-                            Navigator.pushNamed(
-                              context,
-                              ArticleScreen.routeName,
-                              arguments: state.articles![index],
-                            );
-                          },
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              ImageContainer(
-                                width: MediaQuery.of(context).size.width * 0.5,
-                                imageUrl: state.articles![index].urlToImage ?? '',
-                              ),
-                              const SizedBox(height: 10),
-                              Text(
-                                state.articles![index].title ?? 'Title',
-                                maxLines: 2,
-                                overflow: TextOverflow.ellipsis,
-                                style: Theme.of(context).textTheme.bodyLarge!.copyWith(fontWeight: FontWeight.bold, height: 1.5),
-                              ),
-                              const SizedBox(height: 5),
-                              Text('${DateTime.now().difference(state.articles![index].publishedAt!).inHours} hours ago',
-                                  style: Theme.of(context).textTheme.bodySmall),
-                              const SizedBox(height: 5),
-                              Text(
-                                'by ${state.articles![index].author ?? 'Author'}',
-                                overflow: TextOverflow.ellipsis,
-                                style: Theme.of(context).textTheme.bodySmall,
-                              ),
-                            ],
-                          ),
-                        ),
-                      );
-                    },
-                  ),
-                ),
-              ],
-            ),
-          );
-        }
-      },
-    );
-  }
-}
-
-class _NewsOfTheDay extends StatelessWidget {
-  const _NewsOfTheDay({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<RemoteArticlesBloc, RemoteArticleState>(
-      builder: (_, state) {
-        if (state is RemoteArticlesLoading) {
-          return const Center(
-            child: CircularProgressIndicator(),
-          );
-        } else {
-          final article = state.articles!.first;
-          return ImageContainer(
-            height: MediaQuery.of(context).size.height * 0.45,
-            width: double.infinity,
-            padding: const EdgeInsets.all(20.0),
-            imageUrl: article.urlToImage ?? '',
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.end,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                CustomTag(
-                  backgroundColor: Colors.grey.withAlpha(150),
-                  children: [
-                    Text(
-                      'News of the Day',
-                      style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                            color: Colors.white,
-                          ),
-                    ),
-                  ],
-                ),
-                const SizedBox(height: 10),
-                Text(
-                  article.title ?? 'Title',
-                  style:
-                      Theme.of(context).textTheme.headlineSmall!.copyWith(fontWeight: FontWeight.bold, height: 1.25, color: Colors.white),
-                ),
-                TextButton(
-                  onPressed: () {},
-                  style: TextButton.styleFrom(padding: EdgeInsets.zero),
-                  child: Row(
-                    children: [
-                      Text(
-                        'Learn More',
-                        style: Theme.of(context).textTheme.bodyLarge!.copyWith(
-                              color: Colors.white,
-                            ),
-                      ),
-                      const SizedBox(width: 10),
-                      const Icon(
-                        Icons.arrow_right_alt,
-                        color: Colors.white,
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          );
-        }
-      },
     );
   }
 }
